+++ README.md
@@ -37,13 +37,17 @@
         - [x] 기본 11점
         - [x] 총합이 21 초과가 되면 1로 계산한다 (21이 안넘을 때까지 하나씩 1로 계산)
 
-- [ ] Player 
-  - [x] 1 개 받는다
-  - 반복 (n 이 들어오거나 점수 합이 22 이상이 될때까지)
-    - [ ] 현재 가지고 있는 카드의 점수를 계산한다 
-    - [ ] 점수의 합이 22 미만이면 더 받을지의 여부를 입력받는다
-    - [ ] y 면 카드를 새로 한장 받는다
-  - [ ] n 이면 차례가 끝난다
+- [ ] Player
+    - [x] 1 개 받는다
+    - 반복 (n 이 들어오거나 점수 합이 22 이상이 될때까지)
+        - [x] 현재 가지고 있는 카드의 점수를 계산해서 더 받을 수 있는지 여부를 확인
+        - [ ] 점수의 합이 22 미만이면 더 받을지의 여부를 입력받는다
+        - [ ] y 면 카드를 새로 한장 받는다
+    - [ ] n 이면 차례가 끝난다
+- [ ] Dealer
+    - [ ] 딜러가 16 이하면 카드를 계속 받는다
+- [x] Player 와 Dealer 중복되는 코드 제거
+  - [x] Participant 클래스 상속
 
 ## 우아한테크코스 코드리뷰
