+++ src/main/java/blackjack/domain/Player.java
@@ -1,35 +1,10 @@
 package blackjack.domain;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-public class Player {
-    private static final int MAX_SCORE_TO_RECEIVE = 21;
-
-    List<Card> cards;
-    Score score;
+public class Player extends Participant {
 
     public Player(List<Card> cards) {
-        this.cards = cards;
-        this.score = new Score();
-    }
-
-    private List<TrumpNumber> extractNumbers() {
-        return cards.stream()
-                .map(Card::getTrumpNumber)
-                .collect(Collectors.toList());
-    }
-
-    public void receiveCard(Card card) {
-        cards.add(card);
-    }
-
-    public boolean isAbleToReceive() {
-        score.calculateScore(extractNumbers());
-        return score.getScore() <= MAX_SCORE_TO_RECEIVE;
-    }
-
-    public List<Card> getCards() {
-        return cards;
+        super(cards);
     }
 }
